Apply more CloudForms-like documentation styling
With this commit we introduce couple of styling changes to
better match the existing CloudForms documentation styling.
We've also added a developer's README.md to guide documentation
writers with some hints.

Also, with this commit we introduce asciidoc variables to unify
some strings across the whole documentation, please see README.md
for more details.

In the README.md we also describe the PDF building process. Until now
we were just writing documentation in text editor with limited preview,
but it's not good enough - one needs to see the actual PDF in order to
further improve the documentation.

Finally, we also removed some unused files from this repository.
Keeping them in the repo was really confusing. Namely, we removed:

- README.adoc
- SUMMARY.adoc
- vcd_installation_guid/README.adoc

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>