Escalate product title variables to caller
Until now we were using internal variables to refer to product (CloudForms),
but it's better to let caller set those variables to decide whether to use
ManageIQ or CloudForms product title. A list of caller variables can be seen
here:

https://github.com/ManageIQ/manageiq_docs/tree/master/common/attributes

With this commit we slightly modify internal variables in master.adoc to
now make use of those escalated ones.

Fixes: https://github.com/ManageIQ/manageiq-vcloud-docs/issues/6

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>